Add files via upload
Many businesses require a fleet of vehicles to conduct their business. Eventually, all of those vehicles need to be refueled so many of these companies choose to issue credit cards that are assigned to a particular vehicle in the fleet. While this is convenient, it can result in unscrupulous behavior..
Using the provided datasets to flag transactions that require further review.

The Transactions dataset contains Fleet card numbers, Dates, and the amount of the transaction.
The Card Pairings dataset contains the vehicle's ID number and the Fleet card assigned to that vehicle. 
The Fuel Type table contains the vehicle descriptions of fuel type and fuel capacity for ranges of vehicle IDs. 
The Fuel Prices dataset is a listing of average fuel prices by date for regular gasoline and diesel.